Folder_All/requirements.txt
# ==============================================================
# Chatbot Kinh Te Viet Nam - Python Dependencies
# ==============================================================
# Cai dat: pip install -r requirements.txt
# Luu y: Kich hoat .venv truoc khi cai dat
# ==============================================================

# --- Core API & Server ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0

# --- Bao mat & Xac thuc (Security) ---
PyJWT>=2.8.0
httpx>=0.27.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# --- AI Engine: LangChain & LangGraph ---
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
langgraph>=0.2.0

# --- Vector Store & Embeddings ---
langchain-chroma>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0

# --- Data Processing ---
langchain-text-splitters>=0.3.0
numpy>=1.26.0
scikit-learn>=1.5.0
scipy>=1.13.0
tiktoken>=0.7.0

# --- Document Parser (DOCX/PDF to MD) ---
markitdown>=0.1.0
pandas>=2.2.0
openpyxl>=3.1.0

# --- Logging & Monitoring ---
# (Dung thu vien built-in logging cua Python, khong can cai them)

# --- Evaluation ---
ir_datasets>=0.5.0

# --- Frontend (Streamlit - optional) ---
streamlit>=1.38.0
